ping and info updates
