+++ src/commands/Info.ts
@@ -1,6 +1,5 @@
 import { Category, NotBot } from "@discordx/utilities";
-import { oneLine } from "common-tags";
-import { CommandInteraction, Message, MessageActionRow, MessageButton, MessageEmbed } from "discord.js";
+import { CommandInteraction, MessageActionRow, MessageButton, MessageEmbed } from "discord.js";
 import { Client, Discord, Guard, Slash } from "discordx";
 import process from "process";
 
@@ -10,21 +9,7 @@ import { InteractionUtils, ObjectUtil } from "../utils/Utils.js";
 
 @Discord()
 @Category("Utility")
-export class Util {
-    @Slash("ping", {
-        description: "Checks the AvBot's ping to the Discord server"
-    })
-    private async ping(interaction: CommandInteraction, client: Client): Promise<void> {
-        const msg = (await interaction.reply({ content: "Pinging...", fetchReply: true })) as Message;
-        const content = oneLine`
-          ${msg.inGuild() ? `${msg.author},` : ""}
-          Pong! The message round-trip took
-          ${msg.createdTimestamp - interaction.createdTimestamp}ms.
-          ${client.ws.ping ? `The heartbeat ping is ${Math.round(client.ws.ping)}ms.` : ""}
-        `;
-        msg.edit(content);
-    }
-
+export class Info {
     @Slash("info", {
         description: "Provides information about AvBot, and links for adding the bot and joining the support server"
     })
